relocate the oneOfAnyOfType import so that it does not cause a conflict with the message system tsconfig
