+++ packages/fast-tooling-react/src/data-utilities/schema.ts
@@ -1,0 +1,4 @@
+export enum oneOfAnyOfType {
+    anyOf = "anyOf",
+    oneOf = "oneOf",
+}
