+++ packages/fast-tooling-react/src/form/utilities/form.spec.ts
@@ -4,7 +4,8 @@ import { getInitialOneOfAnyOfState, getSchemaByDataLocation } from "./form";
 import mergedOneOfSchema from "../../__tests__/schemas/merged-one-of.schema.json";
 import childrenSchema from "../../__tests__/schemas/children.schema.json";
 import textFieldSchema from "../../__tests__/schemas/textarea.schema.json";
-import { InitialOneOfAnyOfState, oneOfAnyOfType } from "../form-section.props";
+import { InitialOneOfAnyOfState } from "../form-section.props";
+import { oneOfAnyOfType } from "../../data-utilities/schema";
 
 /**
  * Gets a schema by data location (lodash path syntax)
